+++ diesel/src/connection/mod.rs
@@ -136,18 +136,7 @@ pub trait Connection: SimpleConnection + Sized + Send {
     fn execute(&self, query: &str) -> QueryResult<usize>;
 
     #[doc(hidden)]
-    fn query_one<T, U>(&self, source: T) -> QueryResult<U> where
-        T: AsQuery,
-        T::Query: QueryFragment<Self::Backend> + QueryId,
-        Self::Backend: HasSqlType<T::SqlType>,
-        U: Queryable<T::SqlType, Self::Backend>,
-    {
-        self.query_all(source)
-            .and_then(|e: Vec<U>| e.into_iter().next().ok_or(Error::NotFound))
-    }
-
-    #[doc(hidden)]
-    fn query_all<T, U>(&self, source: T) -> QueryResult<Vec<U>> where
+    fn query_by_index<T, U>(&self, source: T) -> QueryResult<Vec<U>> where
         T: AsQuery,
         T::Query: QueryFragment<Self::Backend> + QueryId,
         Self::Backend: HasSqlType<T::SqlType>,