+++ CHANGELOG.md
@@ -31,6 +31,9 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 * `postfix_predicate!` and `postfix_expression!` have been renamed to
   `diesel_postfix_operator!`.
 
+* Trait bounds along the lines of `T: LoadDsl<Conn>, U: Queryable<T::SqlType,
+  Conn::Backend>` should be changed to `T: LoadQuery<Conn, U>`.
+
 ## [0.13.0] - 2017-05-15
 
 ### Added